+++ src/bl4ckscor3/bot/bl4ckb0t/ConfigurationFile.java
@@ -48,6 +48,7 @@ public class ConfigurationFile
 			"select=true",
 			"source=true",
 			"stop=true",
+			"thumbnail=true",
 			"trello=true",
 			"tv=true",
 			"tw=true",