+++ src/bl4ckscor3/bot/bl4ckb0t/Core.java
@@ -20,7 +20,7 @@ public class Core
 	public static Bot bot;
 	private static boolean wasStartedAsWIP;
 	private static final String botName = "bl4ckb0t";
-	private static final String version = "5.8";
+	private static final String version = "5.9";
 	private static ConfigurationFile customConfig;
 
 	public static void main(String args[]) throws IOException, IrcException